original dBase code
This code has been created in 1996. It has been developed for about 5
years over few dBase versions. The last version has been written in
Visual dBase. Don’t have licence anymore and it is not compatible to
valid dBase versions. This code is uploaded to GitHub for anyone who is
interested in. The program has been sold a few times to companies in
machine construction. The program was thought as a data base for
measurement devices or tools. It alarms if a calibration of a
measurement tool has been expired. This is an integral part of an ISO
9001 certification of many companies. So it is a proper program which
needs a code upgrade to live further. It is possible to read DBF files
with free DBF viewers. The win help file can’t be read anymore under
windows 10, since windows doesn’t support old hlp file format anymore
(in win XP it was still possible).